+++ package.json
@@ -14,7 +14,7 @@
     "validateSchemas": "node ./tools/SchemaValidation/SchemaValidationRunner.js",
     "compareSchemas": "node ./tools/SchemaDifference/compareSchemas.js",
     "validateJsonSchemas": "node ./tools/packages/schemaJsonValidationRunner.js",
-    "iModelSchemaValidation": "node ./tools/SchemaValidation/iModelSchemaValidation.js",
+    "iModelSchemaValidation": "node ./tools/SchemaValidation/iModelSchemaValidationRunner.js",
     "genInventory": "node ./tools/packages/generateInventory.js --BisRoot .",
     "genPackages": "node ./tools/packages/generatePackages.js --inventory ./SchemaInventory.json --skipList ./ignoreSchemaList.json --outDir ./packageOut --template ./tools/packages/package.json.template",
     "testPkgGen": "mocha tools/packages/test/**.test.js",
